+++ scalding/src/main/scala/commbank/coppersmith/scalding/FeatureSink.scala
@@ -20,6 +20,7 @@ import com.twitter.util.Encoder
 
 import org.apache.hadoop.fs.Path
 
+import scalaz.Scalaz._
 import scalaz.NonEmptyList
 import scalaz.std.list.listInstance
 import scalaz.syntax.std.list.ToListOpsFromList
@@ -29,9 +30,7 @@ import org.joda.time.DateTime
 
 import au.com.cba.omnia.maestro.api._, Maestro._
 
-import commbank.coppersmith.Feature._
-import commbank.coppersmith.FeatureValue
-
+import commbank.coppersmith._, Feature._
 
 import Partitions.PathComponents
 import HiveSupport.{DelimiterConflictStrategy, FailJob}
@@ -43,7 +42,8 @@ trait FeatureSink {
     * Persist feature values, returning the list of paths written (for committing at the
     * end of the job) or an error if trying to write to a path that is already committed
     */
-  def write(features: TypedPipe[(FeatureValue[Value], FeatureTime)]): WriteResult
+  def write(features: TypedPipe[(FeatureValue[Value], FeatureTime)],
+            metadataSet: List[MetadataSet[Any]]): WriteResult
 }
 
 object FeatureSink {
@@ -135,7 +135,8 @@ case class HiveTextSink[
   delimiter: String = HiveTextSink.Delimiter,
   dcs:       DelimiterConflictStrategy[T] = FailJob[T]()
 ) extends FeatureSink {
-  def write(features: TypedPipe[(FeatureValue[Value], FeatureTime)]) = {
+  def write(features: TypedPipe[(FeatureValue[Value], FeatureTime)],
+            metadataSets: List[MetadataSet[Any]]) = {
     val textPipe = features.map(implicitly[FeatureValueEnc[T]].encode)
 
     val hiveConfig =
@@ -152,6 +153,21 @@ case class HiveTextSink[
     // Note: This needs to be explicitly specified so that the TupleSetter.singleSetter
     // instance isn't used (causing a failure at runtime).
     implicit val tupleSetter: TupleSetter[partition.P] = partition.tupleSetter
-    HiveSupport.writeTextTable(hiveConfig, textPipe)
+
+    val metadataOutput = MetadataOutput.Json1
+    val metadata = metadataOutput.stringify(metadataOutput.doOutput(metadataSets, Set()))
+    val metadataFileName = s"_${metadataSets.map(_.name).mkString("_")}_METADATA.json"
+
+    val result = HiveSupport.writeTextTable(hiveConfig, textPipe)
+    result.flatMap {
+      case Left(x) => Execution.from(Left(x))
+      case Right(ps : Set[Path]) => {
+        val writes: Execution[List[Unit]] = ps.map { p =>
+          val f = new Path(p, metadataFileName)
+          Execution.fromHdfs(Hdfs.write(f, metadata))
+        }.toList.sequence
+        writes.map(_ => Right(ps))
+      }
+    }
   }
 }